factor out tmp dir
